+++ lib/io/CStreamBuffer.cpp
@@ -1,6 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2002 Chris Schoeneman, Nick Bolton, Sorin Sbarnea
+ * Copyright (C) 2002 Chris Schoeneman
  * 
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -10,9 +10,6 @@
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 #include "CStreamBuffer.h"
@@ -79,7 +76,7 @@ CStreamBuffer::pop(UInt32 n)
 	ChunkList::iterator scan = m_chunks.begin();
 	assert(scan != m_chunks.end());
 	while (scan->size() - m_headUsed <= n) {
-		n         -= (UInt32)scan->size() - m_headUsed;
+		n         -= scan->size() - m_headUsed;
 		m_headUsed = 0;
 		scan       = m_chunks.erase(scan);
 		assert(scan != m_chunks.end());
@@ -121,7 +118,7 @@ CStreamBuffer::write(const void* vdata, UInt32 n)
 	while (n > 0) {
 		// choose number of bytes for next chunk
 		assert(scan->size() <= kChunkSize);
-		UInt32 count = kChunkSize - (UInt32)scan->size();
+		UInt32 count = kChunkSize - scan->size();
 		if (count > n)
 			count = n;
