+++ lib/platform/CMSWindowsClipboardAnyTextConverter.cpp
@@ -1,6 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2002 Chris Schoeneman, Nick Bolton, Sorin Sbarnea
+ * Copyright (C) 2002 Chris Schoeneman
  * 
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -10,9 +10,6 @@
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 #include "CMSWindowsClipboardAnyTextConverter.h"
@@ -42,7 +39,7 @@ CMSWindowsClipboardAnyTextConverter::fromIClipboard(const CString& data) const
 {
 	// convert linefeeds and then convert to desired encoding
 	CString text = doFromIClipboard(convertLinefeedToWin32(data));
-	UInt32 size  = (UInt32)text.size();
+	UInt32 size  = text.size();
 
 	// copy to memory handle
 	HGLOBAL gData = GlobalAlloc(GMEM_MOVEABLE | GMEM_DDESHARE, size);
@@ -90,7 +87,7 @@ CMSWindowsClipboardAnyTextConverter::convertLinefeedToWin32(
 
 	// count newlines in string
 	UInt32 numNewlines = 0;
-	UInt32 n = (UInt32)src.size();
+	UInt32 n           = src.size();
 	for (const char* scan = src.c_str(); n > 0; ++scan, --n) {
 		if (*scan == '\n') {
 			++numNewlines;
@@ -105,7 +102,7 @@ CMSWindowsClipboardAnyTextConverter::convertLinefeedToWin32(
 	dst.reserve(src.size() + numNewlines);
 
 	// copy string, converting newlines
-	n = (UInt32)src.size();
+	n = src.size();
 	for (const char* scan = src.c_str(); n > 0; ++scan, --n) {
 		if (scan[0] == '\n') {
 			dst += '\r';
@@ -122,7 +119,7 @@ CMSWindowsClipboardAnyTextConverter::convertLinefeedToUnix(
 {
 	// count newlines in string
 	UInt32 numNewlines = 0;
-	UInt32 n = (UInt32)src.size();
+	UInt32 n           = src.size();
 	for (const char* scan = src.c_str(); n > 0; ++scan, --n) {
 		if (scan[0] == '\r' && scan[1] == '\n') {
 			++numNewlines;
@@ -137,7 +134,7 @@ CMSWindowsClipboardAnyTextConverter::convertLinefeedToUnix(
 	dst.reserve(src.size());
 
 	// copy string, converting newlines
-	n = (UInt32)src.size();
+	n = src.size();
 	for (const char* scan = src.c_str(); n > 0; ++scan, --n) {
 		if (scan[0] != '\r' || scan[1] != '\n') {
 			dst += scan[0];
