+++ lib/server/CClientProxy1_0.cpp
@@ -1,6 +1,6 @@
 /*
  * synergy -- mouse and keyboard sharing utility
- * Copyright (C) 2002 Chris Schoeneman, Nick Bolton, Sorin Sbarnea
+ * Copyright (C) 2002 Chris Schoeneman
  * 
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -10,9 +10,6 @@
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 #include "CClientProxy1_0.h"
@@ -382,7 +379,7 @@ CClientProxy1_0::setOptions(const COptionsList& options)
 	CProtocolUtil::writef(getStream(), kMsgDSetOptions, &options);
 
 	// check options
-	for (UInt32 i = 0, n = (UInt32)options.size(); i < n; i += 2) {
+	for (UInt32 i = 0, n = options.size(); i < n; i += 2) {
 		if (options[i] == kOptionHeartbeat) {
 			double rate = 1.0e-3 * static_cast<double>(options[i + 1]);
 			if (rate <= 0.0) {
